Arduino: Fix bug that truncated laser stim
Deactivate laser will no longer turn off laser

The original intent of the doing this was to ensure that variable-duration
laser stimuli are turned off when the session is finished. However,
this was not the right fix as it caused laser stimulations to be
prematurely terminated.

The right solution is to implement a "max-duration timeout" for the
variable duration laser stim.